Extract AppState
